Complete cog based refactoring
