+++ .gitignore
@@ -1,4 +1,4 @@
 __pycache__
 *.vscode*
-*schedule*
+*.png*
 *launch.json*
